QOL improvements for TypeScript users
