+++ .npmignore
@@ -1,3 +1,5 @@
+# Should be identical to .gitignore with the exception of types/ as we want to publish the types to npm
+
 # Logs
 logs
 *.log
@@ -35,9 +37,6 @@ yarn-error.log*
 
 # eslint
 .eslintcache
-
-# generated types
-types
 
 # IDE gumpf
 .idea
