+++ package.json
@@ -2,12 +2,12 @@
   "name": "gcp-structured-logger",
   "version": "1.4.6",
   "description": "Structured logger for GCP logging",
-  "main": "index.js",
+  "main": "src/index.js",
   "files": [
-    "src/**/*.js",
-    "index.d.ts"
+    "src",
+    "types"
   ],
-  "types": "index.d.ts",
+  "types": "types",
   "repository": {
     "type": "git",
     "url": "https://github.com/bookcreator/gcp-structured-logger"
@@ -24,23 +24,31 @@
     "node": ">=14"
   },
   "scripts": {
+    "build": "tsc -p jsconfig.json",
     "lint": "eslint --cache --fix .",
     "test-debug": "mocha --inspect-brk=50000 --no-timeout",
     "test": "c8 --reporter=html --reporter=text-summary mocha --enable-source-maps",
     "preci-test": "yarn lint",
     "ci-test": "c8 --reporter=text --reporter=text-summary mocha --forbid-only --enable-source-maps",
+    "preversion": "yarn build",
+    "prepublish": "yarn build",
     "postversion": "git push --follow-tags"
   },
-  "dependencies": {},
+  "dependencies": {
+  },
   "devDependencies": {
+    "@google-cloud/logging": "^11.1.0",
     "@types/express-serve-static-core": "^4.17.33",
+    "@typescript-eslint/eslint-plugin": "^7.16.0",
+    "@typescript-eslint/parser": "^7.16.0",
     "c8": "^7.13.0",
     "chai": "^4.3.7",
     "eslint": "^8.38.0",
-    "eslint-plugin-mocha": "^10.0.5",
-    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-mocha": "^10.4.3",
+    "eslint-plugin-n": "^17.9.0",
     "mocha": "^10.2.0",
-    "next": "^12.3.4",
-    "sinon": "^15.0.3"
+    "next": "^14.2.5",
+    "sinon": "^15.0.3",
+    "typescript": "^5.5.3"
   }
 }
