+++ test/StructuredLogger.test.js
@@ -1,3 +1,4 @@
+/* eslint-disable n/global-require */
 const { assert } = require('chai')
 const { createSandbox, match: sinonMatch } = require('sinon')
 const { LogSeverity } = require('..')
@@ -11,6 +12,7 @@ const SERVICE_CONTEXT = {
 
 const SUPPORTS_NATIVE_ERROR_CAUSE = (() => {
    const cause = {}
+   // eslint-disable-next-line n/no-unsupported-features/es-syntax
    return new Error('', { cause }).cause === cause
 })()
 
@@ -21,15 +23,19 @@ describe('StructuredLogger', function () {
    let NODE_ENV
    /** @type {sinon.SinonFakeTimers} */
    let fakeTimers
+
    before(function () {
       loggers = require('../src/StructuredLogger')
    })
+
    after(function () {
       sinon.restore()
    })
+
    beforeEach(function () {
       ({ NODE_ENV } = process.env)
    })
+
    afterEach(function () {
       if (NODE_ENV === undefined) {
          delete process.env.NODE_ENV
@@ -162,10 +168,12 @@ describe('StructuredLogger', function () {
       let logger
       /** @type {sinon.SinonSpy} */
       let writeSpy
+
       before(function () {
          logger = new loggers.StructuredLogger(projectId, logName, SERVICE_CONTEXT, null, null)
          writeSpy = sinon.spy(logger, '_write')
       })
+
       beforeEach(function () {
          writeSpy.resetHistory()
          logger._times.clear()
@@ -200,6 +208,7 @@ describe('StructuredLogger', function () {
 
             /** @type {bigint} */
             let start
+
             beforeEach(function () {
                start = process.hrtime.bigint()
                logger._times.set(LABEL, start)
@@ -703,6 +712,7 @@ describe('StructuredLogger', function () {
 
          it('should use include errors cause', function () {
             const cause = new Error('CAUSE')
+            // eslint-disable-next-line n/no-unsupported-features/es-syntax
             const error = SUPPORTS_NATIVE_ERROR_CAUSE ? new Error('TOP LEVEL', { cause }) : (() => {
                const e = new Error('TOP LEVEL')
                e.cause = cause
