+++ test/cleanup-for-json.test.js
@@ -1,3 +1,4 @@
+/* eslint-disable n/global-require */
 const { inspect } = require('util')
 const { assert } = require('chai')
 const sinon = require('sinon')
@@ -6,6 +7,7 @@ const cleanupForJSON = require('../src/cleanup-for-json')
 
 const SUPPORTS_NATIVE_ERROR_CAUSE = (() => {
    const cause = {}
+   // eslint-disable-next-line n/no-unsupported-features/es-syntax
    return new Error('', { cause }).cause === cause
 })()
 
@@ -38,6 +40,7 @@ describe('cleanup-for-json', function () {
             flags: 'gi',
          })
       })
+
       it('should serialise constructed to object', function () {
          const v = new RegExp('abcde\\(f', 'mig')
          assert.deepStrictEqual(cleanupForJSON(v), {
@@ -52,11 +55,13 @@ describe('cleanup-for-json', function () {
       it('should serialise to array with same order', function () {
          assert.sameDeepMembers(cleanupForJSON([1, 2, '3', false, true, 1.2]), [1, 2, '3', false, true, 1.2])
       })
+
       it('should serialise contents with same order', function () {
          const data = require('crypto').randomBytes(25)
          const date = new Date()
          assert.sameDeepMembers(cleanupForJSON([data, /abcde\(f/ig, date]), [{ '@type': 'Buffer', length: data.length, base64: data.toString('base64') }, { '@type': 'RegExp', source: 'abcde\\(f', flags: 'gi' }, date.toISOString()])
       })
+
       it('should replace circular references', function () {
          const v = [123]
          v.push(v)
@@ -68,11 +73,13 @@ describe('cleanup-for-json', function () {
       it('should serialise to array', function () {
          assert.deepStrictEqual(cleanupForJSON(new Set([1, 2, '3', false, true, 1.2])), [1, 2, '3', false, true, 1.2])
       })
+
       it('should serialise contents', function () {
          const data = require('crypto').randomBytes(25)
          const date = new Date()
          assert.deepStrictEqual(cleanupForJSON(new Set([data, /abcde\(f/ig, date])), [{ '@type': 'Buffer', length: data.length, base64: data.toString('base64') }, { '@type': 'RegExp', source: 'abcde\\(f', flags: 'gi' }, date.toISOString()])
       })
+
       it('should replace circular references', function () {
          const v = new Set([123])
          v.add(v)
@@ -92,6 +99,7 @@ describe('cleanup-for-json', function () {
             { key: 'hello', value: false },
          ])
       })
+
       it('should serialise values', function () {
          const data = require('crypto').randomBytes(25)
          const date = new Date()
@@ -105,6 +113,7 @@ describe('cleanup-for-json', function () {
             { key: 'hello', value: date.toISOString() },
          ])
       })
+
       it('should serialise keys', function () {
          const data = require('crypto').randomBytes(25)
          const date = new Date()
@@ -118,6 +127,7 @@ describe('cleanup-for-json', function () {
             { key: date.toISOString(), value: 'hello' },
          ])
       })
+
       it('should replace value circular references', function () {
          const v = new Map([['123', 123]])
          v.set('map', v)
@@ -126,6 +136,7 @@ describe('cleanup-for-json', function () {
             { key: 'map', value: '[Circular]' }
          ])
       })
+
       it('should replace key circular references', function () {
          const v = new Map([['123', 123]])
          v.set(v, 'map')
@@ -146,6 +157,7 @@ describe('cleanup-for-json', function () {
             cause: undefined,
          })
       })
+
       it('should serialise to object (Error subclass)', function () {
          const e = new TypeError('Some error')
          assert.deepStrictEqual(cleanupForJSON(e), {
@@ -155,6 +167,7 @@ describe('cleanup-for-json', function () {
             cause: undefined,
          })
       })
+
       it('should serialise to object (Error prototype)', function () {
          const e = Object.create(Error.prototype)
          assert.deepStrictEqual(cleanupForJSON(e), {
@@ -164,6 +177,7 @@ describe('cleanup-for-json', function () {
             cause: undefined,
          })
       })
+
       it('should serialise to object (custom Error subclass)', function () {
          class CustomError extends Error { }
          const e = new CustomError('Some error')
@@ -174,6 +188,7 @@ describe('cleanup-for-json', function () {
             cause: undefined,
          })
       })
+
       it('should serialise nested Error to object', function () {
          const err1 = new Error('Some error 1')
          const err2 = new Error('Some error 2')
@@ -210,6 +225,7 @@ describe('cleanup-for-json', function () {
 
       it('should serialise Errors cause', function () {
          const cause = new Error('CAUSE')
+         // eslint-disable-next-line n/no-unsupported-features/es-syntax
          const error = SUPPORTS_NATIVE_ERROR_CAUSE ? new Error('TOP LEVEL', { cause }) : (() => {
             const e = new Error('TOP LEVEL')
             e.cause = cause
@@ -284,6 +300,7 @@ describe('cleanup-for-json', function () {
          }
          assert.deepStrictEqual(cleanupForJSON(v), v)
       })
+
       it('should serialise values', function () {
          const data = require('crypto').randomBytes(25)
          const date = new Date()
@@ -298,6 +315,7 @@ describe('cleanup-for-json', function () {
             'hello': date.toISOString(),
          })
       })
+
       it('should serialise object with Symbol.iterator function', function () {
          const v = {
             [Symbol.iterator]() {
@@ -315,6 +333,7 @@ describe('cleanup-for-json', function () {
          }
          assert.deepStrictEqual(cleanupForJSON(v), ['hello', 'world'])
       })
+
       it('should serialise object with Symbol.iterator generator', function () {
          const v = {
             *[Symbol.iterator]() {
@@ -324,6 +343,7 @@ describe('cleanup-for-json', function () {
          }
          assert.deepStrictEqual(cleanupForJSON(v), ['hello', 'world'])
       })
+
       it('should serialise to object using toJSON', function () {
          const v = {
             toJSON() {
@@ -332,6 +352,7 @@ describe('cleanup-for-json', function () {
          }
          assert.deepStrictEqual(cleanupForJSON(v), { hello: 'world' })
       })
+
       it('should serialise to object passing root key toJSON', function () {
          const v = {
             toJSON() {
@@ -344,6 +365,7 @@ describe('cleanup-for-json', function () {
 
          sinon.assert.calledOnceWithExactly(toJSON, '')
       })
+
       it('should serialise to object passing parent key toJSON', function () {
          const v = {
             key: {
@@ -358,6 +380,7 @@ describe('cleanup-for-json', function () {
 
          sinon.assert.calledOnceWithExactly(toJSON, 'key')
       })
+
       it('should serialise to object passing parent key index toJSON', function () {
          const v = [
             {
@@ -372,6 +395,7 @@ describe('cleanup-for-json', function () {
 
          sinon.assert.calledOnceWithExactly(toJSON, '0')
       })
+
       it('should serialise to object passing map parent key toJSON', function () {
          const obj = {
             toJSON() {
@@ -385,6 +409,7 @@ describe('cleanup-for-json', function () {
 
          sinon.assert.calledOnceWithExactly(toJSON, '123')
       })
+
       it('should serialise to object passing map key toJSON', function () {
          const obj = {
             toJSON() {
@@ -398,6 +423,7 @@ describe('cleanup-for-json', function () {
 
          sinon.assert.calledOnceWithExactly(toJSON, 'map')
       })
+
       it('should serialise to object using toJSON that returns array', function () {
          const v = {
             toJSON() {
@@ -406,6 +432,7 @@ describe('cleanup-for-json', function () {
          }
          assert.deepStrictEqual(cleanupForJSON(v), [{ hello: 'world' }])
       })
+
       it('should serialise to object using toJSON that returns string', function () {
          const v = {
             toJSON() {
@@ -414,6 +441,7 @@ describe('cleanup-for-json', function () {
          }
          assert.deepStrictEqual(cleanupForJSON(v), 'hello')
       })
+
       it('should serialise to object using toJSON that returns null', function () {
          const v = new Map()
          v.set('key', {
@@ -423,6 +451,7 @@ describe('cleanup-for-json', function () {
          })
          assert.deepStrictEqual(cleanupForJSON(v), [{ key: 'key', value: null }])
       })
+
       it('should serialise to object using inspect.custom', function () {
          const v = {
             [inspect.custom]() {
@@ -431,6 +460,7 @@ describe('cleanup-for-json', function () {
          }
          assert.deepStrictEqual(cleanupForJSON(v), { hello: 'world' })
       })
+
       it('should serialise to object using inspect.custom passing default inspect options', function () {
          const v = {
             [inspect.custom]() {
@@ -443,6 +473,7 @@ describe('cleanup-for-json', function () {
 
          sinon.assert.calledOnceWithExactly(customInspect, inspect.defaultOptions.depth, inspect.defaultOptions)
       })
+
       it('should serialise to object using inspect.custom that returns array', function () {
          const v = {
             [inspect.custom]() {
@@ -451,6 +482,7 @@ describe('cleanup-for-json', function () {
          }
          assert.deepStrictEqual(cleanupForJSON(v), [{ hello: 'world' }])
       })
+
       it('should serialise to object using inspect.custom that returns string', function () {
          const v = {
             [inspect.custom]() {
@@ -459,6 +491,7 @@ describe('cleanup-for-json', function () {
          }
          assert.deepStrictEqual(cleanupForJSON(v), 'hello')
       })
+
       it('should serialise to object using inspect.custom that returns null', function () {
          const v = new Map()
          v.set('key', {
