+++ test/trace-context.test.js
@@ -16,9 +16,11 @@ describe('trace-context', function () {
    it('should ignore requests with trace header', function () {
       assert.deepStrictEqual(traceContext(PROJECT_ID, makeReq()), {})
    })
+
    it('should ignore requests with invalid trace header', function () {
       assert.deepStrictEqual(traceContext(PROJECT_ID, makeReq('blah')), {})
    })
+
    it('should extract context for requests with trace header', function () {
       const traceId = '59973d340da5c40f77349df948ef7531'
       const spanId = 15
@@ -27,6 +29,7 @@ describe('trace-context', function () {
          trace: `projects/${PROJECT_ID}/traces/${traceId}`,
       })
    })
+
    it('should extract context for requests with trace header (with options)', function () {
       const traceId = '703ea69e0cf1952e228fe792159d5996'
       const spanId = 9007199254740991
@@ -35,6 +38,7 @@ describe('trace-context', function () {
          trace: `projects/${PROJECT_ID}/traces/${traceId}`,
       })
    })
+
    it('should extract context for requests with trace header (64-bit span)', function () {
       const traceId = 'fb48e517f910e54d5b6f55e16c65fe3a'
       const spanId = '9007199254740992'
@@ -43,6 +47,7 @@ describe('trace-context', function () {
          trace: `projects/${PROJECT_ID}/traces/${traceId}`,
       })
    })
+
    it('should extract context for requests with trace header (64-bit span with options)', function () {
       const traceId = '0eac9cb48de916b2c5d5fb0eafcbc712'
       const spanId = '17710341711684079263'
